config: Add per-container colors
Per-container color schemes allow a number of use-cases to be
implemented, particularly giving visual feedback to the user based on
specific properties of a view. This change adds that functionality in a
minimalistic manner:

We extend the client.<class> commands to accept container context, i.e.
work with criteria. sway_container is extended to carry its own local
border color classes configuration. container_get_current_colors() is
extended to check for custom configuration of the container or its
parents. Otherwise the main config values are used as before. To ease
detection whether a per-container configuration exists for a particular
color class, the configuration structure is switched to pointers
referencing dynamically allocated buffers. To allow consistent
programmatic access to the various color classes, the the configuration
structure is switched to an array indexed using an enum.

Command client.<class> is extended with subcommands "default" to remove
per-container configuration and "global" to explicitly address the
global configuration.

Test plan:

client.focused #000000 #000000 #ffffff #000000 #000000
for_window [app_id="foo"] client.focused #000000 #900000 #ffffff #000000 #000000

Start app foo and notice that the title bar is red when focused.

Signed-off-by: Michael Weiser <michael.weiser@gmx.de>